feat: Add MotiL CSV loader and OTFormer readout variants
- Add MotiLMoleculeCSVDataset and motil_scafford_balance split strategy
- Support disable_rum_ot ablation and readout_variant configs in OTFormer
- Update finetune configs to use MotiL CSV source and pretrained models
- Add ablation experiment configs and run script
- Improve pretraining auto-resume logic

+++ gps/logger.py
@@ -69,7 +69,7 @@ class CustomLogger(Logger):
 
     # task properties
     def classification_binary(self):
-        true = torch.cat(self._true).squeeze(-1)
+        true = torch.cat(self._true).squeeze(-1).long()
         pred_score = torch.cat(self._pred)
         pred_int = self._get_pred_int(pred_score)
